[tool] Add a test
In the prior commit I moved a function (probably-diagram-yaml?) from one
ns to another. In doing so I noticed that we had no test for it. So I
thought I should add one. Of course I thought it’d be no big deal, but
it turned out to not be so simple. Once again the power of generative
testing revealed that I hadn’t thought about edge cases. For example,
what was split-file supposed to do it its input was a blank string?
So some adjustments were required to account for such edge cases.