[tool] Fix a bug in fc4.yaml/split-file spec
In an earlier commit I changed the args spec in the fspec to an or spec
without accounting for the differences in the structure of the conformed
value of each arg. I also hadn’t properly thought through the semantics
of the two cases. Really it boils down to a blank string or a non-blank
string.