EP05 WIP++ Continue refactor of rendering
Specifically, change how fc4.io.render handles errors — instead of
calling fc4.cli.util/fail which generally calls System/exit — it now
just throws an exception and leaves it to the caller to decide what to
do with it. This was needed because when the `edit` command is used from
the CLI and one or more files are being watched for changes, we don’t
want to exit the JVM when there are errors — we just want to print them
out and then continue, keep watching for more changes and then try
again.