[tool EP05] Move fns from fc4.cli.render to fc4.io.render
This is preparatory refactoring because as per EP05 (#111) we’ll soon be
adding a new `edit` CLI command that will be implemented in
`fc4.cli.edit` and will also need to render YAML files. In other words,
the functionality that until now was only in fc4.cli.render will need to
be used by two different namespaces under fc4.cli, so we needed that
functionality to be in a third namespace.

I chose to move these functions out of the `cli` source tree and into
the `main` source tree because by convention we don’t have tests for
the code under `cli` but we do have tests for that under `main`. And
since this code will now be shared by more than one ns, and anyway
since it was fairly extensive, it should be tested.