[tool EP05] Change how fc4.io.render handles errors
Instead of calling fc4.cli.util/fail which generally calls System/exit
it now just throws an exception and leaves it to the caller to decide
what to do with it.

This preparatory refactoring for EP05 (#111) for which we’ll soon be
adding an `edit` subcommand, and when a user has run that from the CLI
and one or more files are being watched for changes, we don’t want to
exit the JVM when there are errors — we just want to print them out and
then continue, keep watching for more changes and then try again.