[tool] Remove obsolete namespace fc4.repl
This is a remnant from back in the day when the tool was only usable
from within a Clojure REPL. That hasn’t been a recommended or useful
way to use the tool for awhile, so it’s time to drop this now-useless
code. Lets us drop a dependency and a now-unused function too.

Closes #117