[tool] Remove obsolete guidance on building an uberjar
This was needed when the tool needed to be compiled on Java 8–10 and
then tested on Java 11. But as of 52b7e6b the tool can now be
compiled successfully on Java 11. So I think this whole section is
no longer needed. If someone wants to manually build an uberjar or
a complete distribution package (including the renderer) they can
check out the scripts in `bin` and/or check out how the packages
get built in CircleCI.