[docs] Convert Markdown inside table to HTML
I used [grip][grip] to preview the markup wherein Markdown was nested
inside HTML, and it rendered just fine. So that situation renders just
fine within github.com, e.g. when viewing the repo or the readme. I had
thought that GitHub Pages used the same Markdown renderer as github.com,
but apparently not — when this was published to [the main FC4 website][home]
the Markdown markup was escaped, not rendered.

Converting the content in the table to HTML should eliminate this problem.
It’s maybe not super-ideal, but table support in Markdown is iffy in
general, so I can live with this. I think it’s an acceptable situation
until we switch the main website to a custom Jekyll theme and I can
start using CSS to lay out the page rather than tables.

[grip]: https://github.com/joeyespo/grip
[home]: https://fundingcircle.github.io/fc4-framework/