[docs] Decomplect website and developer docs
Until now, the entire repo (i.e. the root of the repo) has been the
“publication source” for the repo’s GitHub Pages site build. This was
totally fine for awhile, but it’s started to cause issues.

First, I noticed that because I was using `tool/README.md` as both the
README for devs wanting to work on fc4-tool *and* as the source of the
tool page of the website, I had both end-user docs and developer docs on
the same page, which was sub-optimal — it made both kinds of docs less
distinct and less explicit, and has been less than user-friendly.

Then, through the course of multiple PRs (#142, #143, #144) I realized
that the root-level README.md, which was being rendered as the home
page of the repo on github.com *and* as the home page of the website on
GitHub Pages, was being processed by a different Markdown processor for
each context. This made it tricky to figure out what markup would work
where, and meant that any time I’d change any of the README files I’d
need to preview the changes using both grip *and* Jekyll so as to test
how the markup would be translated to HTML in both contexts.

Well, that seemed untenable, and I’d already been planning to decouple
the website (including the end-user docs) from the dev docs (accessible
via the repo at github.com) and move the website into the `docs` dir,
because I’m working on a bunch of new end-user docs and the root dir of
the repo is starting to get really messy in some of those branches.

So there’s a couple of things happening here:

* All end-user docs have been moved into the new root-level dir `docs`
* The README.md files are now focused on the use case of developers
  viewing them in the context of the repo on github.com — i.e. they’re
  the developer docs
* I added a README.md to `docs` that’s meant to be viewed in the context
  of the repo — this is for developers and/or documentarians who want to
  work on the website