Update CircleCI config to version 2.1
Initially, as per CircleCI’s guidance[1], we need to change just the
version number and run a build to make sure the build succeeds using
the “new processing service” -- whatever that is.

So I guess this is preparatory refactoring to starting to actually use
the new features of the new version.

[1]: https://circleci.com/docs/2.0/reusing-config/#getting-started-with-config-reuse